refactor(dm): extracts openDmForUser
