+++ src/components/views/elements/MessageButton.tsx
@@ -1,26 +1,16 @@
 import React, { useContext, useState } from "react";
 import MatrixClientContext from "matrix-react-sdk/src/contexts/MatrixClientContext";
 import AccessibleButton from "matrix-react-sdk/src/components/views/elements/AccessibleButton";
-import { MatrixClient, RoomMember, User } from "matrix-js-sdk/src/matrix";
-import { DirectoryMember, startDmOnFirstMessage } from "matrix-react-sdk/src/utils/direct-messages";
 import { useAtom } from "jotai";
 
 import { Member } from "../right_panel/UserInfo";
 import { Icon as SendMessage } from "../../../../res/themes/superhero/img/icons/send.svg";
 import { BareUser, botAccountsAtom } from "../../../atoms";
+import { openDmForUser } from "../../../utils";
 
 /**
  * Converts the member to a DirectoryMember and starts a DM with them.
  */
-export async function openDmForUser(matrixClient: MatrixClient, user: Member | BareUser): Promise<void> {
-    const avatarUrl = user instanceof User ? user.avatarUrl : user instanceof RoomMember ? user.getMxcAvatarUrl() : "";
-    const startDmUser = new DirectoryMember({
-        user_id: user.userId,
-        display_name: user.rawDisplayName,
-        avatar_url: avatarUrl,
-    });
-    await startDmOnFirstMessage(matrixClient, [startDmUser]);
-}
 
 export const MessageButton = ({
     member,
